backend/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.1
apscheduler==3.10.4
requests==2.32.3
gunicorn==22.0.0
